UWP update

+++ ManaBob/ManaBob/Models/User.cs
@@ -6,14 +6,13 @@ using System.Threading.Tasks;
 
 namespace ManaBob
 {
-    // - Note
-    //      '사용자'의 기본 클래스
-    //      - ID        : 사용자들의 고유 식별자
-    // - To Do
-    //      ID의 충돌 방지 방안?
+    /// <summary>
+    ///      '사용자'의 기본 클래스
+    ///      - ID    : 사용자들의 고유 식별자. 유효하지 않은 경우 음수를 사용
+    /// </summary>
     public class User
     {
-        public UInt32 ID    { get; set; }
+        public Int64  ID    { get; set; }
         public String Name  { get; set; }
     }
